+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning v2.0.0](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] (2.4.2)
+### Fixed
+- Message archive entries are stored by id in a stable way.
+
 ## [2.4.1] - 2023-07-06
 ### Fixed
 - Restore old content URL on `/content` instead of new `/api/content`